requirements.txt
openai>=1.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
duckduckgo-search>=6.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pyTelegramBotAPI>=4.14.0
